recipe_agent/requirements.txt
google-adk==1.21.0
google-cloud-aiplatform==1.132.0
google-genai==1.56.0
pillow==12.1.0
python-dotenv==1.2.1
reportlab==4.4.7